// front/lib/generated/asset_registry/types/kycStatus.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

export enum KycStatus {
  Pending,
  Approved,
  Revoked,
  Expired,
}

export type KycStatusArgs = KycStatus;

export function getKycStatusEncoder(): FixedSizeEncoder<KycStatusArgs> {
  return getEnumEncoder(KycStatus);
}

export function getKycStatusDecoder(): FixedSizeDecoder<KycStatus> {
  return getEnumDecoder(KycStatus);
}

export function getKycStatusCodec(): FixedSizeCodec<KycStatusArgs, KycStatus> {
  return combineCodec(getKycStatusEncoder(), getKycStatusDecoder());
}