// front/lib/generated/asset_registry/types/realizeAction.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

export enum RealizeAction {
  BurnAndAttest,
  TransferToBeneficiary,
  BurnAndPayout,
}

export type RealizeActionArgs = RealizeAction;

export function getRealizeActionEncoder(): FixedSizeEncoder<RealizeActionArgs> {
  return getEnumEncoder(RealizeAction);
}

export function getRealizeActionDecoder(): FixedSizeDecoder<RealizeAction> {
  return getEnumDecoder(RealizeAction);
}

export function getRealizeActionCodec(): FixedSizeCodec<
  RealizeActionArgs,
  RealizeAction
> {
  return combineCodec(getRealizeActionEncoder(), getRealizeActionDecoder());
}